+++ src/app/componentes/header/header.component.html
@@ -31,7 +31,7 @@
         </li>
 
         <li class="nav">
-          <a class="nav-link " href="#" [routerLink]="['/']">
+          <a class="nav-link " href="#" [routerLink]="['/recursos_comunitarios']">
             Recursos
           </a>
         </li>
