--- src/app/componentes/recurso-comunitario/lista-recursos-comunitarios/lista-recursos-comunitarios.component.html
@@ -1,37 +1,0 @@
-<main class="text-center row p-0 m-0 justify-content-center">
-  <div class="container">
-    <nav aria-label="breadcrumb">
-      <ol class="breadcrumb">
-        <li class="breadcrumb-item"><a [routerLink]="['/inicio']">Inicio</a></li>
-        <li class="breadcrumb-item active" aria-current="page">Recursos comunitarios</li>
-      </ol>
-    </nav>
-    <h1 class="component-title">Recursos comunitarios</h1>
-    <div class="form-group mb-3">
-      <input id="campoBusqueda" type="text" class="form-control" placeholder="Busqueda..." [(ngModel)]="inputBusqueda">
-    </div>
-    <div class="table-responsive m-2">
-      <table id="tableList" class="table table-hover text-center">
-        <thead>
-        <tr class="table-purple">
-          <th class="ordenable" (click)="ordenacionTabla(0, 'string')" scope="col">Nombre <i class="fas fa-sort"></i></th>
-          <th class="ordenable" (click)="ordenacionTabla(1, 'number')" scope="col">Teléfono <i class="fas fa-sort"></i></th>
-          <th class="ordenable" (click)="ordenacionTabla(2, 'string')" scope="col">Tipo de recurso comunitario <i class="fas fa-sort"></i></th>
-          <th class="ordenable" (click)="ordenacionTabla(3, 'string')" scope="col">Dirección <i class="fas fa-sort"></i></th>
-          <th scope="col"></th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr *ngFor="let recurso_comunitario of recursos_comunitarios | filtroBusqueda: inputBusqueda | paginate: { itemsPerPage: 10, currentPage: numPaginacion }"
-            app-item-resurso-comunitario [recurso_comunitario]="recurso_comunitario">
-        </tr>
-        </tbody>
-      </table>
-      <pagination-controls
-        previousLabel="Anterior"
-        nextLabel="Siguiente"
-        responsive="true"
-        (pageChange)="numPaginacion = $event"></pagination-controls>
-    </div>
-  </div>
-</main>
