--- src/app/componentes/recursos/recursos.component.html
@@ -1,17 +1,0 @@
-<table id="tableList" class="table table-hover text-center">
-  <thead>
-  <tr class="table-purple">
-    <th>Nombre <i class="fas fa-sort"></i></th>
-    <th>Teléfono <i class="fas fa-sort"></i></th>
-    <th>Localidad <i class="fas fa-sort"></i></th>
-    <th>Provincia <i class="fas fa-sort"></i></th>
-    <th>Direccion <i class="fas fa-sort"></i></th>
-    <th>CP <i class="fas fa-sort"></i></th>
-    <th>PAC <i class="fas fa-sort"></i></th>
-  </tr>
-  </thead>
-  <tbody>
-  <tr> <!-- Hay que hacer un bucle for con los datos de los recursos -->
-  </tr>
-  </tbody>
-</table>
