--- src/app/componentes/recursos/recursos.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-recursos',
-  templateUrl: './recursos.component.html',
-  styleUrls: ['./recursos.component.scss']
-})
-export class RecursosComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
