+++ src/app/components/recursos/recurso-comunitario/crear-recurso-comunitario/crear-recurso-comunitario.component.scss
@@ -1,0 +1,3 @@
+.no-valido{
+  border: 2px solid #f10734;
+}
