+++ src/app/components/recursos/recurso-comunitario/crear-recurso-comunitario/crear-recurso-comunitario.component.ts
@@ -1,15 +1,15 @@
 import {Component, OnInit} from '@angular/core';
-import {IDireccion} from '../../../interfaces/i-direccion';
-import {IRecursoComunitario} from '../../../interfaces/i-recurso-comunitario';
-import {ITipoRecursoComunitario} from '../../../interfaces/i-tipo-recurso-comunitario';
+import {IDireccion} from '../../../../interfaces/i-direccion';
+import {IRecursoComunitario} from '../../../../interfaces/i-recurso-comunitario';
+import {ITipoRecursoComunitario} from '../../../../interfaces/i-tipo-recurso-comunitario';
 import {Title} from '@angular/platform-browser';
 import {ActivatedRoute, Router} from '@angular/router';
-import {CargaDireccionService} from '../../../servicios/carga-direccion.service';
-import {CargaRecursoComunitarioService} from '../../../servicios/carga-recurso-comunitario.service';
-import {Direccion} from '../../../clases/direccion';
-import {RecursoComunitario} from '../../../clases/recurso-comunitario';
+import {CargaDireccionService} from '../../../../servicios/carga-direccion.service';
+import {CargaRecursoComunitarioService} from '../../../../services/recursos/carga-recurso-comunitario.service';
+import {Direccion} from '../../../../clases/direccion';
+import {RecursoComunitario} from '../../../../clases/recurso-comunitario';
 import Swal from "sweetalert2";
-import {environment} from "../../../../environments/environment";
+import {environment} from "../../../../../environments/environment";
 
 
 @Component({
