+++ src/app/components/recursos/recurso-comunitario/item-resurso-comunitario/item-resurso-comunitario.component.html
@@ -1,0 +1,17 @@
+<td>{{ recurso_comunitario.nombre }}</td>
+<td>{{ recurso_comunitario.telefono }}</td>
+<td>{{ recurso_comunitario.id_tipos_recurso_comunitario.nombre }}</td>
+<td>{{ recurso_comunitario.id_direccion.localidad }}</td>
+<td>{{ recurso_comunitario.id_direccion.provincia}}</td>
+<td>{{ recurso_comunitario.id_direccion.codigo_postal}}</td>
+<td>{{ recurso_comunitario.id_direccion.direccion}}</td>
+<td>
+  <button type="submit" class="btn btn-outline-secondary btn-modify"
+          [routerLink]="['/recursos_comunitarios/modificar', recurso_comunitario.id]"> <!-- Este routerLink nos llevará a la página de modificación -->
+    <i class="fa fa-pen"></i>
+  </button>
+  <button class="btn  btn-outline-secondary m-1"><i class="fa fa-eye" style="color: #abb8ce;"></i></button>
+  <!-- Botón para borrar el recurso comunitario -->
+  <button (click)="eliminarRecurso()" type="submit" class="btn btn-delete m-1"><i class="fa fa-trash"></i></button>
+</td>
+
