+++ src/app/components/recursos/recurso-comunitario/item-resurso-comunitario/item-resurso-comunitario.component.spec.ts
