+++ src/app/components/recursos/recurso-comunitario/modificar-recurso-comunitario/modificar-recurso-comunitario.component.ts
@@ -1,13 +1,13 @@
 import {Component, OnInit} from '@angular/core';
-import {IDireccion} from '../../../interfaces/i-direccion';
-import {IRecursoComunitario} from '../../../interfaces/i-recurso-comunitario';
-import {ITipoRecursoComunitario} from '../../../interfaces/i-tipo-recurso-comunitario';
+import {IDireccion} from '../../../../interfaces/i-direccion';
+import {IRecursoComunitario} from '../../../../interfaces/i-recurso-comunitario';
+import {ITipoRecursoComunitario} from '../../../../interfaces/i-tipo-recurso-comunitario';
 import {ActivatedRoute, Router} from '@angular/router';
 import {Title} from '@angular/platform-browser';
-import {CargaDireccionService} from '../../../servicios/carga-direccion.service';
-import {CargaRecursoComunitarioService} from '../../../servicios/carga-recurso-comunitario.service';
+import {CargaDireccionService} from '../../../../servicios/carga-direccion.service';
+import {CargaRecursoComunitarioService} from '../../../../services/recursos/carga-recurso-comunitario.service';
 import Swal from "sweetalert2";
-import {environment} from "../../../../environments/environment";
+import {environment} from "../../../../../environments/environment";
 
 
 @Component({
