+++ src/app/components/recursos/tipo-recurso-comunitario/crear-tipo-recurso-comunitario/crear-tipo-recurso-comunitario.component.ts
@@ -1,33 +1,32 @@
 import {Component, OnInit} from '@angular/core';
-import {ITipoRecursoComunitario} from '../../../interfaces/i-tipo-recurso-comunitario';
-import {ActivatedRoute, Router} from '@angular/router';
+import {ITipoRecursoComunitario} from '../../../../interfaces/i-tipo-recurso-comunitario';
 import {Title} from '@angular/platform-browser';
-import {CargaTipoRecursoComunitarioService} from '../../../servicios/carga-tipo-recurso-comunitario.service';
+import {ActivatedRoute, Router} from '@angular/router';
+import {CargaTipoRecursoComunitarioService} from '../../../../services/recursos/carga-tipo-recurso-comunitario.service';
+import {TipoRecursoComunitario} from '../../../../clases/tipo-recurso-comunitario';
 import Swal from "sweetalert2";
-import {environment} from "../../../../environments/environment";
+import {environment} from "../../../../../environments/environment";
 
 
 @Component({
-  selector: 'app-modificar-tipo-recurso-comunitario',
-  templateUrl: './modificar-tipo-recurso-comunitario.component.html',
-  styleUrls: ['./modificar-tipo-recurso-comunitario.component.scss']
+  selector: 'app-crear-tipo-recurso-comunitario',
+  templateUrl: './crear-tipo-recurso-comunitario.component.html',
+  styleUrls: ['./crear-tipo-recurso-comunitario.component.scss']
 })
 
-export class ModificarTipoRecursoComunitarioComponent implements OnInit {
+export class CrearTipoRecursoComunitarioComponent implements OnInit {
   public tipo_recurso_comunitario: ITipoRecursoComunitario;
-  public idTipoRecursoComunitario: number;
 
-  constructor(private route: ActivatedRoute, private titleService: Title, private cargaTiposRecursosComunitarios: CargaTipoRecursoComunitarioService, private router: Router) {
+  constructor(private titleService: Title, private route: ActivatedRoute, private cargaTiposRecursosComunitarios: CargaTipoRecursoComunitarioService, private router: Router) {
   }
 
   ngOnInit(): void {
-    this.tipo_recurso_comunitario = this.route.snapshot.data['tipo_recurso_comunitario'];
-    this.idTipoRecursoComunitario = this.route.snapshot.params['id'];
-    this.titleService.setTitle('Modificar tipo recurso comunitario ' + this.idTipoRecursoComunitario);
+    this.titleService.setTitle('Nuevo tipo recurso comunitario');
+    this.tipo_recurso_comunitario = new TipoRecursoComunitario();
   }
 
-  modificarTipoRecursoComunitario(): void {
-    this.cargaTiposRecursosComunitarios.modificarTipoRecursoComunitario(this.tipo_recurso_comunitario).subscribe(
+  nuevoTipoRecursoComunitario(): void {
+    this.cargaTiposRecursosComunitarios.nuevoTipoRecursoComunitario(this.tipo_recurso_comunitario).subscribe(
       e => {
         this.alertExito()
         this.router.navigate(['/tipos_recursos_comunitarios']);
@@ -37,7 +36,7 @@ export class ModificarTipoRecursoComunitarioComponent implements OnInit {
       }
     );
   }
-//Toast para el Alert indicando que la operación fue exitosa
+  //Toast para el Alert indicando que la operación fue exitosa
   alertExito() :void {
     const Toast = Swal.mixin({
       toast: true,
@@ -54,7 +53,7 @@ export class ModificarTipoRecursoComunitarioComponent implements OnInit {
 
     Toast.fire({
       icon: 'success',
-      title: environment.fraseModificar,
+      title: environment.fraseCrear,
     })
   }
   //Toast para el alert indicando que hubo algún error en la operación
@@ -73,8 +72,7 @@ export class ModificarTipoRecursoComunitarioComponent implements OnInit {
 
     Toast.fire({
       icon: 'error',
-      title: environment.fraseErrorModificar
+      title: environment.fraseErrorCrear
     })
   }
-
 }
