+++ src/app/components/recursos/tipo-recurso-comunitario/item-tipo-recurso-comunitario/item-tipo-recurso-comunitario.component.ts
@@ -1,28 +1,24 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {IRecursoComunitario} from '../../../interfaces/i-recurso-comunitario';
-
+import {ITipoRecursoComunitario} from '../../../../interfaces/i-tipo-recurso-comunitario';
 import Swal from "sweetalert2";
-import {CargaTipoRecursoComunitarioService} from "../../../servicios/carga-tipo-recurso-comunitario.service";
-import {CargaRecursoComunitarioService} from "../../../servicios/carga-recurso-comunitario.service";
-import {environment} from "../../../../environments/environment";
+import {CargaTipoRecursoComunitarioService} from "../../../../services/recursos/carga-tipo-recurso-comunitario.service";
+import {environment} from "../../../../../environments/environment";
 import {Router} from "@angular/router";
 
-
 @Component({
-  selector: 'app-item-resurso-comunitario, [app-item-resurso-comunitario]',
-  templateUrl: './item-resurso-comunitario.component.html',
-  styleUrls: ['./item-resurso-comunitario.component.scss']
+  selector: 'app-item-tipo-recurso-comunitario, [app-item-tipo-recurso-comunitario]',
+  templateUrl: './item-tipo-recurso-comunitario.component.html',
+  styleUrls: ['./item-tipo-recurso-comunitario.component.scss']
 })
 
-export class ItemResursoComunitarioComponent implements OnInit {
-  @Input() public recurso_comunitario: IRecursoComunitario;
+export class ItemTipoRecursoComunitarioComponent implements OnInit {
+  @Input() public tipo_recurso_comunitario: ITipoRecursoComunitario;
 
-  constructor(private cargaRecursoComunitario: CargaRecursoComunitarioService, private router:Router) {
+  constructor(private cargaTipoRecursoComunitario: CargaTipoRecursoComunitarioService, private router: Router) {
   }
 
   ngOnInit(): void {
   }
-
   //Toast para el Alert indicando que la operación fue exitosa
   alertExito() :void {
     const Toast = Swal.mixin({
@@ -62,22 +58,24 @@ export class ItemResursoComunitarioComponent implements OnInit {
       title: environment.fraseErrorEliminar
     })
   }
+
+
   modalConfirmacion(): void {
     Swal.fire({
-      title: '¿Está seguro que desea eliminar este recurso comunitario?',
+      title: '¿Está seguro que desea eliminar este tipo de recurso comunitario?',
       showCancelButton: true,
       confirmButtonText: 'Aceptar',
     }).then((result) => {
       /* Read more about isConfirmed, isDenied below */
       if (result.isConfirmed) {
-        this.eliminarRecursoComunitario('recursos_comunitarios')
+        this.eliminarTipoRecursoComunitario('tipos_recursos_comunitarios')
       }
     })
   }
-  eliminarRecursoComunitario(ruta:string) : void{
-    this.cargaRecursoComunitario.eliminarRecursoComunitario(this.recurso_comunitario).subscribe(
+  eliminarTipoRecursoComunitario(ruta: string) : void{
+    this.cargaTipoRecursoComunitario.eliminarTipoRecursoComunitario(this.tipo_recurso_comunitario).subscribe(
       e=>{
-        this.router.navigateByUrl(ruta+'/borrado/'+this.recurso_comunitario.id, {skipLocationChange: true}).then(() => {
+        this.router.navigateByUrl(ruta+'/borrado/'+this.tipo_recurso_comunitario.id, {skipLocationChange: true}).then(() => {
           this.router.navigate([ruta]);
         });
         //Si el elemento se ha borrado con exito, llama al método que muestra el alert de Exito
@@ -89,4 +87,5 @@ export class ItemResursoComunitarioComponent implements OnInit {
       }
     )
   }
+
 }
