+++ src/app/interfaces/i-direccion.ts
@@ -1,4 +1,4 @@
-export interface IDireccion {
+export interface IDireccion { // Con esta interfaz identificaremos cual es la direccion
   id: number;
   localidad: string;
   provincia: string;
