+++ src/app/interfaces/i-recurso-comunitario.ts
@@ -5,6 +5,7 @@ export interface IRecursoComunitario {
   id: number;
   nombre: string;
   telefono: string;
+  // Necesitamos otras interfaces para completar el recurso comunitario
   id_tipos_recurso_comunitario: ITipoRecursoComunitario;
   id_direccion: IDireccion;
 }
