+++ src/app/interfaces/i-tipo-recurso-comunitario.ts
@@ -1,4 +1,4 @@
-export interface ITipoRecursoComunitario {
+export interface ITipoRecursoComunitario { // Con esta interfaz identificaremos cual es el tipo de recursos comunitario
   id: any;
   nombre: string;
 }
