+++ src/app/services/recursos/carga-tipo-recurso-comunitario.service.ts
@@ -1,8 +1,8 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {Observable} from 'rxjs';
-import {ITipoRecursoComunitario} from '../interfaces/i-tipo-recurso-comunitario';
-import {environment} from "../../environments/environment";
+import {ITipoRecursoComunitario} from '../../interfaces/i-tipo-recurso-comunitario';
+import {environment} from "../../../environments/environment";
 
 @Injectable({
   providedIn: 'root'
